refactor(internal): move pkg to repo root
