+++ internal/doc.go
@@ -5,7 +5,12 @@
 // Package internal provides a collection of internal
 // packages used for the Vela runtime environments.
 //
+// More information:
+//
+//   * https://golang.org/doc/go1.4#internalpackages
+//   * https://docs.google.com/document/d/1e8kOo3r51b2BWtTs_1uADIA5djfXhPT36s6eHVRIvaU/edit
+//
 // Usage:
 //
-// 	import "github.com/go-vela/pkg-runtime/runtime/internal"
+// 	import "github.com/go-vela/pkg-runtime/internal"
 package internal
