+++ internal/image/image_test.go
