+++ internal/internal.go
@@ -1,0 +1,10 @@
+// Copyright (c) 2020 Target Brands, Inc. All rights reserved.
+//
+// Use of this source code is governed by the LICENSE file in this repository.
+
+package internal
+
+// For more information:
+//
+//   * https://golang.org/doc/go1.4#internalpackages
+//   * https://docs.google.com/document/d/1e8kOo3r51b2BWtTs_1uADIA5djfXhPT36s6eHVRIvaU/edit
