--- runtime/internal/internal.go
@@ -1,5 +1,0 @@
-// Copyright (c) 2020 Target Brands, Inc. All rights reserved.
-//
-// Use of this source code is governed by the LICENSE file in this repository.
-
-package internal
