+++ runtime/kubernetes/container.go
@@ -13,7 +13,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/go-vela/pkg-runtime/runtime/internal/image"
+	"github.com/go-vela/pkg-runtime/internal/image"
 	"github.com/go-vela/types/pipeline"
 
 	"github.com/sirupsen/logrus"
